the-one-drivered/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4